modularizacion - qualificationAdapter.js
